ia/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

googlesearch-python==1.2.3
gTTS==2.4.0
gym==0.26.2
markovify==0.2.3
openai==0.28.1
pygame==2.5.2
SpeechRecognition==3.10.0
sympy==1.12
transformers==4.19.2
